Changed Imagebar to have a center image, larger than the others.
The as an image slides into the center, it smoothly transitions into a bigger size.
Also removed the extra four images in the imagebar that were there to make it slide.
Also tweaked slick settings a bit.
Also made the scrollbar with noscript not appear if the images could all fit into the viewport.